pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "attention-research"
version = "0.1.0"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.11"
dependencies = [
  "torch>=2.1.0",
  "torchvision>=0.16.0",
  "torchaudio>=2.1.0",
  "jax[cuda12_pip]>=0.4.20",
  "jaxlib>=0.4.20",
  "flax>=0.7.5",
  "optax>=0.1.7",
  "transformers>=4.35.0",
  "datasets>=2.14.0",
  "tokenizers>=0.14.0",
  "numpy>=1.24.0",
  "scipy>=1.11.0",
  "matplotlib>=3.7.0",
  "seaborn>=0.12.0",
  "pandas>=2.0.0",
  "tqdm>=4.65.0",
  "wandb>=0.16.0",
  "python-dotenv>=1.0.0",
  "omegaconf>=2.3.0",
  "hydra-core>=1.3.0",
  "tensorboard>=2.14.0",
  "rich>=13.0.0",
  "psutil>=5.9.0",
  "pynvml>=11.5.0",
]
description = "Research repository for comparing attention mechanisms in PyTorch and JAX"
authors = [{ name = "Attention Research Team" }]
keywords = ["attention", "transformers", "pytorch", "jax", "machine-learning"]
classifiers = [
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

# [project.urls]
# Documentation = "https://link.to.documentation"

[tool.setuptools.packages.find]
where = ["src"]
namespaces = true

[dependency-groups]
dev = [
  "mypy",
  "pre-commit",
  "pytest",
  "pytest-cov",
  "pytest-mock",
  "ruff",
]

# [project.scripts]
# script-name = "module:function"

# [tool.setuptools.package-data]
# template = ["py.typed"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [ # See https://docs.astral.sh/ruff/rules/
  "F",    # flake8 (pyflakes)
  "E",    # pycodestyle errors
  "W",    # pycodestyle warnings
  "I",    # isort
  "N",    # pep8-naming
  "D",    # flake8-docstrings (pydocstyle)
  "UP",   # pyupgrade
  "S",    # flake8-bandit
  "BLE",  # flake8-blind-except
  "FBT",  # flake8-boolean-trap maybe add this
  "B",    # flake8-bugbear
  "A",    # flake8-builtins
  "C4",   # flake8-comprehensions
  "DTZ",  # flake8-datetimez
  "T10",  # flake8-debugger (to remove breakpoint())
  "EM",   # flake8-errmsg
  "G",    # flake8-logging-format
  "INP",  # flake8-no-pep420
  "PIE",  # flake8-pie
  "PT",   # flake8-pytest-style
  "Q",    # flake8-quotes
  "RSE",  # flake8-raise (to remove unused parethesis in raise)
  "RET",  # flake8-return
  "SLF",  # flake8-self
  "SLOT", # flake8-slots (enforce __slots__ for some subclasses)
  "SIM",  # flake8-simplify
  "TID",  # flask8-tidy-imports
  "TCH",  # flake8-type-checking
  "PTH",  # flake8-use-pathlib
  "PGH",  # pygrep-hooks (mostly to check for invalid-mock-access)
  "TRY",  # tryceratops
  "NPY",  # numpy-specific rules
  "PERF", # perflint
  "FURB", # refurb
  "LOG",  # flake8-logging
  "RUF",  # ruff-specific rules
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "SLF", "INP"]
"scripts/**/*.py" = ["INP001", "T20"]

[tool.mypy]
python_version = "3.11"
disable_error_code = ["import-untyped"]
no_implicit_optional = true
check_untyped_defs = true
strict_equality = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
addopts = "--cov=src/ -v --cov-report=term-missing --durations=20 --no-cov-on-fail"
